pyproject.toml
[project]
name = "qa-automation-browser-automation"
version = "0.1.0"
description = "QA Automation for Shopify with MCP and NLP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.9.5",
    "mcp>=1.0.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
